Render buttons in card footer if width is avobe narrow mode threshold
